+++ ccmt-cff-rest-api/chalicelib/util/formSubmit/util.py
@@ -154,6 +154,8 @@ def calculate_price(expressionString, data):
 def format_payment(total, currency='USD'):
     if currency == "USD":
         return "${:,.2f}".format(total)
+    if currency == "INR":
+        return "₹{:,.2f}".format(total)
     return "{} {:,.2f}".format(currency, total)
 def format_paymentInfo(paymentInfo):
     return format_payment(paymentInfo.get("total", "N/A"), paymentInfo.get("currency", "USD"))